requirements.txt
# Web UI Framework
gradio==5.49.1

# AWS Services
boto3==1.40.59

# Environment Configuration
python-dotenv==1.1.1

# LlamaIndex Core & Components
llama-index==0.12.52
llama-index-core>=0.10.0
llama-index-retrievers-bedrock==0.3.0

# LLM Integration
llama-index-llms-gemini>=0.1.0
google-generativeai>=0.4.0

# Optional: For enhanced performance
aiohttp>=3.9.0
httpx>=0.25.0